docs-prompts-server/requirements.txt
mcp>=0.1.0
PyYAML>=6.0
watchdog>=3.0.0
sentence-transformers>=2.2.2
jinja2>=3.1.0
gitpython>=3.1.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
